Refactor PIR sensor to individual module.
