+++ src/modules/blockly/generators/propc/sensors.js
@@ -227,125 +227,6 @@ Blockly.propc.joystick_input_yaxis = function() {
 Blockly.Blocks.joystick_input_xaxis = Blockly.Blocks.joystick_input_yaxis;
 Blockly.propc.joystick_input_xaxis = Blockly.propc.joystick_input_yaxis;
 
-// ---------------- PIR Sensor Blocks --------------------------------
-
-/**
- *
- * @type {{
- *  init: Blockly.Blocks.PIR_Sensor.init,
- *  addPinMenu: *,
- *  mutationToDom: *,
- *  helpUrl: string,
- *  setToOther: *,
- *  domToMutation: *
- * }}
- */
-Blockly.Blocks.PIR_Sensor = {
-  helpUrl: Blockly.MSG_PIR_HELPURL,
-  init: function() {
-    this.setTooltip(Blockly.MSG_PIR_SENSOR_TOOLTIP);
-    this.setColour(colorPalette.getColor('input'));
-    this.pinChoices = ['PIN'];
-    this.otherPin = [false];
-    this.addPinMenu('PIR sensor PIN', null, 0);
-    this.setInputsInline(true);
-    this.setNextStatement(false, null);
-    this.setPreviousStatement(false, null);
-    this.setOutput(true, 'Number');
-  },
-
-  // mutationToDom: Blockly.Blocks['sensor_ping'].mutationToDom,
-  mutationToDom: function() {
-    const container = document.createElement('mutation');
-    for (let pinOpt = 0; pinOpt < this.pinChoices.length; pinOpt++) {
-      // TODO: verify that otherPin is in fact a boolean
-      container.setAttribute('otherpin' +
-          pinOpt, this.otherPin[pinOpt].toString());
-    }
-    return container;
-  },
-
-  // domToMutation: Blockly.Blocks['sensor_ping'].domToMutation,
-  domToMutation: function(xmlElement) {
-    for (let pinOpt = 0; pinOpt < this.pinChoices.length; pinOpt++) {
-      const op = xmlElement.getAttribute('otherpin' + pinOpt);
-      if (op === 'true') {
-        this.setToOther('other', this.moveBefore, pinOpt);
-      }
-    }
-  },
-
-  // addPinMenu: Blockly.Blocks['sensor_ping'].addPinMenu,
-  /**
-   * Add a pin to the block configuration
-   * @param {string} label
-   * @param {boolean} moveBefore
-   * @param {number} pinOpt is an index into the pins array
-   */
-  addPinMenu: function(label, moveBefore, pinOpt) {
-    const profile = getDefaultProfile();
-    this.appendDummyInput('SET_PIN')
-        .appendField(label, 'LABEL')
-        .appendField(new Blockly.FieldDropdown(
-            profile.digital.concat([['other', 'other']]),
-            function(op) {
-              // eslint-disable-next-line no-invalid-this
-              this.getSourceBlock().setToOther(op, moveBefore, pinOpt);
-            }), this.pinChoices[pinOpt]);
-    this.moveBefore = moveBefore;
-    this.otherPin[pinOpt] = false;
-  },
-
-  // setToOther: Blockly.Blocks['sensor_ping'].setToOther,
-  /**
-   * Set the pin to default 'other' as defined in the board type profile
-   * @param {string} op
-   * @param {boolean} moveBefore
-   * @param {number} pinOpt
-   */
-  setToOther: function(op, moveBefore, pinOpt) {
-    if (op === 'other') {
-      const profile = getDefaultProfile();
-      this.otherPin[pinOpt] = true;
-      const label = this.getFieldValue('LABEL');
-      if (this.getInput('SET_PIN')) {
-        this.removeInput('SET_PIN');
-      }
-      this.appendValueInput(this.pinChoices[pinOpt])
-          .appendField(label)
-          .setCheck('Number')
-          .appendRange('A,' + profile.digital.toString());
-      if (moveBefore) {
-        this.moveInputBefore(this.pinChoices[pinOpt], moveBefore);
-      } else {
-        const currBlockTimeout = this;
-        setTimeout(function() {
-          currBlockTimeout.render();
-        }, 200);
-      }
-    }
-  },
-
-};
-
-/**
- *
- * @return {[string, number]}
- * @constructor
- */
-Blockly.propc.PIR_Sensor = function() {
-  let pin = '0';
-
-  if (this.otherPin[0]) {
-    pin = Blockly.propc.valueToCode(
-        this, this.pinChoices[0], Blockly.propc.ORDER_ATOMIC) || '0';
-  } else {
-    pin = this.getFieldValue(this.pinChoices[0]);
-  }
-  return ['input(' + pin + ')', Blockly.propc.ORDER_ATOMIC];
-};
-
-
 // ---------------- Sound Impact Sensor Blocks -----------------------
 
 /**
